requirements.txt
anyio==3.4.0
asgiref==3.4.1
attrs==21.4.0
beautifulsoup4==4.10.0
cachetools==4.2.4
certifi==2021.10.8
charset-normalizer==2.0.10
click==8.0.3
docutils==0.18.1
fastapi==0.71.0
flit_core==3.6.0
google==3.0.0
google-api-core==2.3.2
google-auth==2.3.3
google-cloud==0.34.0
google-cloud-pubsub==2.9.0
googleapis-common-protos==1.54.0
grpc-google-iam-v1==0.12.3
grpcio==1.43.0
grpcio-status==1.43.0
h11==0.12.0
httptools==0.3.0
idna==3.3
iniconfig==1.1.1
libcst==0.3.23
mypy-extensions==0.4.3
packaging==21.3
parameterized==0.8.1
pluggy==1.0.0
proto-plus==1.19.8
protobuf==3.19.1
py==1.11.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydantic==1.9.0
PyJWT==2.3.0
pyparsing==3.0.6
pytest==6.2.5
python-dotenv==0.19.2
PyYAML==6.0
requests==2.27.0
rsa==4.8
six==1.16.0
slack-sdk==3.13.0
sniffio==1.2.0
soupsieve==2.3.1
starlette==0.17.1
toml==0.10.2
tomli==2.0.0
tomli_w==1.0.0
typing-inspect==0.7.1
typing_extensions==4.0.1
urllib3==1.26.7
uvicorn==0.16.0
uvloop==0.16.0
watchgod==0.7
websockets==10.1
